Финальное задание 14 спринта

+++ src/main/java/ru/practicum/shareit/item/storage/InMemoryItemStorage.java
@@ -10,10 +10,10 @@ import java.util.Map;
 @Component
 public class InMemoryItemStorage implements ItemStorage {
 
-    private int uniqueId = 0;
-    public final Map<Integer, Item> items = new HashMap<>();
+    private Long uniqueId = Long.valueOf(0);
+    public final Map<Long, Item> items = new HashMap<>();
 
-    private int getUniqueId() {
+    private Long getUniqueId() {
         uniqueId++;
         return uniqueId;
     }
@@ -36,7 +36,7 @@ public class InMemoryItemStorage implements ItemStorage {
     }
 
     @Override
-    public Item getItemById(int id) {
+    public Item getItemById(Long id) {
         return items.get(id);
     }
